requirements.txt
flask>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pandas>=1.3.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
pillow>=8.0.0
werkzeug>=2.0.0
flask-socketio>=5.0.0
eventlet>=0.30.0
gunicorn>=20.1.0
gevent>=21.1.0
supervisor>=4.2.0
lxml>=4.6.0
python-engineio>=4.8.0
python-socketio>=5.10.0
python-dotenv>=1.0.0
flask-cli>=0.4.0
moviepy>=1.0.3
openai-whisper>=20231117
torch>=2.1.0
numpy>=1.24.0
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.4
google-generativeai>=0.3.0

# Thêm thư viện cho chức năng nâng cao chất lượng ảnh
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
opencv-python>=4.5.0

selenium
webdriver-manager
chromedriver-autoinstaller


